requirements.txt
Flask
Flask-Bootstrap4